Added Daily Budget controller

+++ src/main/java/com/djajkoski/budgeteer/repository/BudgetRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface BudgetRepository extends CrudRepository<DailyBudget, Integer> {
+public interface BudgetRepository extends CrudRepository<DailyBudget, Long> {
 }
